Saturday Jan 15 Work
- ParamEditors all get a Mixin base class hoding the ref and stuff
- Rebind of combo box swaps the contents properly
- Re-structure the DataInterfaces.h into wrappers/juce/data
- InteractionData restructured. Checks at runtime for correctness
  and includes the id for self documentation
