+++ wrappers/juce/data/SCXTData.h
@@ -1,0 +1,14 @@
+//
+// Created by Paul Walker on 1/4/22.
+//
+
+#ifndef SHORTCIRCUIT_SCXTDATA_H
+#define SHORTCIRCUIT_SCXTDATA_H
+
+#include "data/ParameterProxy.h"
+#include "data/BasicTypes.h"
+#include "data/UIStateProxy.h"
+#include "data/DataModel.h"
+#include "data/DataApplicators.h"
+
+#endif // SHORTCIRCUIT_SCXTDATA_H
