+++ wrappers/juce/data/UIStateProxy.h
@@ -1,0 +1,146 @@
+//
+// Created by Paul Walker on 1/15/22.
+//
+
+#ifndef SHORTCIRCUIT_UISTATEPROXY_H
+#define SHORTCIRCUIT_UISTATEPROXY_H
+
+#include "sampler_wrapper_actiondata.h"
+#include "data/BasicTypes.h"
+
+namespace scxt
+{
+namespace data
+{
+
+/*
+ * The UIStateProxy is a class which handles messages and keeps an appropriate state.
+ * Components can render it for bulk action. For instance there woudl be a ZoneMapProxy
+ * and so on. Each UIStateProxy registerred with the editor gets all the messages.
+ */
+class UIStateProxy
+{
+  public:
+    struct Invalidatable
+    {
+        virtual ~Invalidatable() = default;
+        virtual void onProxyUpdate() = 0;
+    };
+
+    struct InvalidateAndRepaintGuard
+    {
+        InvalidateAndRepaintGuard(UIStateProxy &p) : proxy(p), go{true} {}
+        ~InvalidateAndRepaintGuard()
+        {
+            if (go)
+                proxy.markNeedsRepaintAndProxyUpdate();
+        }
+        bool deactivate()
+        {
+            go = false;
+            return go;
+        }
+        UIStateProxy &proxy;
+        bool go;
+    };
+
+    virtual ~UIStateProxy() = default;
+    virtual bool processActionData(const actiondata &d) = 0;
+    std::unordered_set<juce::Component *> clients;
+    void repaintClients() const
+    {
+        for (auto c : clients)
+        {
+            c->repaint();
+        }
+    }
+
+    enum Validity
+    {
+        VALID = 0,
+        NEEDS_REPAINT = 1 << 1,
+        NEEDS_PROXY_UPDATE = 1 << 2
+    };
+    uint32_t validity{VALID};
+
+    void markNeedsRepaint() { validity = validity | NEEDS_REPAINT; }
+    void markNeedsProxyUpdate() { validity = validity | NEEDS_PROXY_UPDATE; }
+    void markNeedsRepaintAndProxyUpdate()
+    {
+        validity = validity | NEEDS_PROXY_UPDATE | NEEDS_REPAINT;
+    }
+
+    void sweepValidity()
+    {
+        if (validity == VALID)
+            return;
+        bool rp = validity & NEEDS_REPAINT;
+        bool px = validity & NEEDS_PROXY_UPDATE;
+        validity = VALID;
+        for (auto c : clients)
+        {
+            if (rp)
+            {
+                c->repaint();
+            }
+
+            if (px)
+            {
+                if (auto q = dynamic_cast<Invalidatable *>(c))
+                {
+                    q->onProxyUpdate();
+                }
+            }
+        }
+    }
+
+    bool collectStringEntries(const actiondata &ad, NameList &v)
+    {
+        if (std::holds_alternative<VAction>(ad.actiontype))
+        {
+            auto at = std::get<VAction>(ad.actiontype);
+            if (at == vga_entry_clearall)
+            {
+                v.data.clear();
+                v.update_count++;
+                return true;
+            }
+            if (at == vga_entry_add_ival_from_self)
+            {
+                v.data.push_back(ad.data.str);
+                v.update_count++;
+                return true;
+            }
+            if (at == vga_entry_add_ival_from_self_with_id)
+            {
+                auto entry = ad.data.i[0];
+                auto val = (char *)(&ad.data.str[4]);
+                if (entry >= v.data.size())
+                    v.data.resize(entry + 1);
+                v.data[entry] = val;
+                v.update_count++;
+                return true;
+            }
+            if (at == vga_entry_replace_label_on_id)
+            {
+                auto entry = ad.data.i[0];
+                auto val = (char *)(&ad.data.str[4]);
+                v.data[entry] = val;
+                v.update_count++;
+
+                return true;
+            }
+        }
+        return false;
+    }
+    bool collectStringEntries(const actiondata &ad, InteractionId id, NameList &v)
+    {
+        if (ad.id != id)
+            return false;
+        return collectStringEntries(ad, v);
+    }
+};
+
+} // namespace data
+} // namespace scxt
+#endif // SHORTCIRCUIT_UISTATEPROXY_H
